REFACTOR: ListBox 컴포넌트 리팩토링
- ComboMenu로 컴포넌트명 변경
- 태그 변경
- 스타일 관련 props 간소화
- 스타일 믹스인 사용
- 이벤트, 노드 타입 변경
